+++ src/SeablastConstant.php
@@ -6,6 +6,7 @@ namespace Seablast\Seablast;
 
 /**
  * @api
+ * Each string MUST start with SB to avoid unintended value collision
  */
 class SeablastConstant
 {
@@ -13,6 +14,14 @@ class SeablastConstant
      * @var string Running or under construction
      */
     public const FLAG_WEB_RUNNING = 'SB:web:running';
+    /**
+     * @var string Redirection database should be looked up (unused so far)
+     */
+    public const FLAG_CHECK_REDIRECTOR = 'SB:redirector:running';
+    /**
+     * @var string Output JSON as HTML instead of application/json so that Tracy is displayed
+     */
+    public const FLAG_DEBUG_JSON = 'SB:debug:json';
     /**
      * @var string Level of PHP built-in error_reporting
      */
@@ -59,29 +68,46 @@ class SeablastConstant
      * TODO: make sure this is needed
      * @var string
      */
-    public const BACKYARD_LOGGING_LEVEL = 'BACKYARD_LOGGING_LEVEL';
+    public const BACKYARD_LOGGING_LEVEL = 'SB:BACKYARD_LOGGING_LEVEL';
     /**
      * @var string flag whether to send emails to admin
      */
-    public const ADMIN_MAIL_ENABLED = 'ADMIN_MAIL:ENABLED';
+    public const ADMIN_MAIL_ENABLED = 'SB:ADMIN_MAIL:ENABLED';
     /**
      * @var string string admin's email address
      */
-    public const ADMIN_MAIL_ADDRESS = 'ADMIN_MAIL:ADDRESS';
+    public const ADMIN_MAIL_ADDRESS = 'SB:ADMIN_MAIL:ADDRESS';
     /**
      * @var string string[] IP addresses where to show Tracy
      */
-    public const DEBUG_IP_LIST = 'DEBUG_IP_LIST';
+    public const DEBUG_IP_LIST = 'SB:DEBUG_IP_LIST';
     /**
      * @var string string[] mapping slugs to templates and tables
      */
-    public const APP_MAPPING = 'APP_MAPPING';
+    public const APP_MAPPING = 'SB:APP_MAPPING';
     /**
      * @var string string with path to directory with Latte templates
      */
-    public const LATTE_TEMPLATE = 'LATTE_TEMPLATE';
+    public const LATTE_TEMPLATE = 'SB:LATTE_TEMPLATE';
     /**
      * @var string string with path to directory with cache for Latte
      */
-    public const LATTE_CACHE = 'LATTE_CACHE';
+    public const LATTE_CACHE = 'SB:LATTE_CACHE';
+    /**
+     * @var string Name of an expected and accepted numeric GET argument
+     */
+    public const SB_GET_ARGUMENT_ID = 'SB_GET_ARGUMENT_ID';
+    /**
+     * @var string Name of an expected and accepted string GET argument
+     */
+    public const SB_GET_ARGUMENT_CODE = 'SB_GET_ARGUMENT_CODE';
+    /**
+     * @var string Int to forced update of external CSS and Javascript files
+     * TODO use this in seablast-dist
+     */
+    public const SB_WEB_FORCE_ASSET_VERSION = 'SB_WEB_FORCE_ASSET_VERSION';
+    /**
+     * @var string The absolute URL of the root of the application
+     */
+    public const SB_APP_ROOT_ABSOLUTE_URL = 'SB_APP_ROOT_ABSOLUTE_URL';
 }