+++ src/SeablastSetup.php
@@ -1,11 +1,8 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Seablast\Seablast;
-
-//use Tracy\Debugger;
-//use Tracy\ILogger;
-
-//use Webmozart\Assert\Assert;
 
 class SeablastSetup
 {